date filter session issue solved for trip search

+++ application/controllers/user.php
@@ -55,7 +55,9 @@ class User extends CI_Controller {
 		$param4=$this->uri->segment(6);
 
 		//unset search session condition before navigation
+		if($param1!='trips'){
 		$this->mysession->delete('condition');
+		}
 
         if($this->session_check()==true) {
 		if($param1==''){
@@ -914,6 +916,7 @@ class User extends CI_Controller {
 	public function Trips($param2){
 		if($this->session_check()==true|| $this->customer_session_check()==true || $this->driver_session_check()==true) 
 		{
+		
 			//pagination first page link setup
 			if($param2=='1'){ $param2='0'; }
 			if($param2==''){ $this->mysession->delete('condition');$param2='0'; }
@@ -945,10 +948,11 @@ class User extends CI_Controller {
 				$where_arry['driver_id']=$_REQUEST['drivers'];
 				$where_arry['trip_status_id']=$_REQUEST['trip_status_id'];
 				$where_arry['customer_group_id']=$_REQUEST['cgroups'];
-				$like_arry['customer_name']=$_REQUEST['customer'];	
+				$like_arry['customer_name']=$_REQUEST['customer'];
+				$this->mysession->set('condition',array("where"=>$where_arry,"like"=>$like_arry));				
 			}
 
-			$this->mysession->set('condition',array("where"=>$where_arry,"like"=>$like_arry));	
+				
 
 			//get session condition
 			if($this->mysession->get('condition')!=''){
